Made the player fall
I updated the code to make the player to fall by 1 pixel every frame, and the background is being rendered every frame now instead of just at the start.
I also defined a updatePlayerRect(Player* player) function so that you can just change the x/y/width/height values and update the rect as to keep all of the values in sync. Could I just use the rect and abandon the x/y/width/height values? Probably. That may come in the future.