Fixed score/stillTime unsyncing issue
stillTime was reset to 0 despite everything above maxStillTime needing to be saved, so taking maxStillTime from stillTime means that stillTime and score will be identical after the decimal point, so the annoyingEnemy is always spawned at 3, 6, 9, 12, 15, etc.  seconds instead of 3, 6.0001, 9.0035, 12.0041, 15.005, etc. (exaggerated for effect). Basically a stupid mistake that was really easy to fix once I noticed it :)